hop-node: logical separation of class logic
