+++ packages/hop-node/src/watchers/classes/ContractBase.ts
@@ -10,7 +10,7 @@ import {
 } from 'src/constants'
 import { Event, PayableOverrides } from '@ethersproject/contracts'
 import { EventEmitter } from 'events'
-import { FinalityService, FinalityStrategyType } from 'src/finality/FinalityService'
+import { FinalityService } from 'src/finality/FinalityService'
 import { config as globalConfig } from 'src/config'
 
 export type TxOverrides = PayableOverrides & {from?: string, value?: BigNumberish}
@@ -36,7 +36,7 @@ export default class ContractBase extends EventEmitter {
     this.finalityService = new FinalityService(
       this.contract.provider,
       this.chainSlug,
-      FinalityStrategyType.Bonder
+      FinalityService.FinalityStrategyType.Bonder
     )
   }
